Adjusted how ParserTargets are stored/referenced; pre-game text parsing
Previously, fetching ParserTargets was a fairly call-heavy process (about 4 nested function calls to get the NPC object) which led to some runtime bloat. Implemented a SpecificParserTarget class which stores a reference to the NPC class on the object, and directly fetches the description stored or delegates to the NPC's description as appropriate. This led to some issues due to the implementation of ParserTarget's members as static and UtilText attempting to retrieve the targets before they were initialized. That case was erroneous, though, since it was trying to parse text that was pre-game load and would never be shown; added a check to ignore such parsing if the game hasn't started yet.