Abstracted Parser engine out of UtilText and proper implementation of GraalVM
The naive implementation of Graal via ScriptEngine was pretty low effort, but functional as a way to 1-to-1 swap-out old Nashorn stuff but it was absurdly slow. Separating the ScriptEngine into a wrapper class lets us delegate the work to whatever underlying JS interpreter we're using at the time (fingers crossed nothing happens to Graal in the short term), and also means doing a proper Graal implementation is pretty trivial.